Move schematracker to pkg/sqlcache/schematracker
